+++ src/ts/Managers/CampaignRefreshManager.ts
@@ -8,11 +8,12 @@ import { Configuration } from 'Models/Configuration';
 import { Diagnostics } from 'Utilities/Diagnostics';
 import { AbstractAdUnit } from 'AdUnits/AbstractAdUnit';
 import { INativeResponse } from 'Utilities/Request';
+import { Session } from 'Models/Session';
+import { SdkStats } from 'Utilities/SdkStats';
 
 export class CampaignRefreshManager {
     public static NoFillDelay = 3600;
     public static ErrorRefillDelay = 3600;
-    public static QuickRefillTestDelay = 60;
 
     private _nativeBridge: NativeBridge;
     private _wakeUpManager: WakeUpManager;
@@ -21,9 +22,6 @@ export class CampaignRefreshManager {
     private _currentAdUnit: AbstractAdUnit<Campaign>;
     private _refillTimestamp: number;
     private _needsRefill = true;
-
-    private _singleCampaignMode: boolean = false;
-    private _singleCampaignErrorCount: number = 0;
 
     constructor(nativeBridge: NativeBridge, wakeUpManager: WakeUpManager, campaignManager: CampaignManager, configuration: Configuration) {
         this._nativeBridge = nativeBridge;
@@ -35,7 +33,7 @@ export class CampaignRefreshManager {
         this._campaignManager.onCampaign.subscribe((placementId, campaign) => this.onCampaign(placementId, campaign));
         this._campaignManager.onNoFill.subscribe(placementId => this.onNoFill(placementId));
         this._campaignManager.onError.subscribe((error, placementIds, rawAdPlan) => this.onError(error, placementIds, rawAdPlan));
-        this._campaignManager.onAdPlanReceived.subscribe((refreshDelay, singleCampaignMode) => this.onAdPlanReceived(refreshDelay, singleCampaignMode));
+        this._campaignManager.onAdPlanReceived.subscribe((refreshDelay) => this.onAdPlanReceived(refreshDelay));
     }
 
     public getCampaign(placementId: string): Campaign | undefined {
@@ -61,7 +59,6 @@ export class CampaignRefreshManager {
             this.setPlacementStates(PlacementState.WAITING, this._configuration.getPlacementIds());
             this._refillTimestamp = 0;
             this.invalidateCampaigns(false, this._configuration.getPlacementIds());
-            this._singleCampaignMode = false;
             return this._campaignManager.request();
         } else if(this.checkForExpiredCampaigns()) {
             return this.onCampaignExpired();
@@ -135,31 +132,26 @@ export class CampaignRefreshManager {
     }
 
     private onCampaign(placementId: string, campaign: Campaign) {
-        this._singleCampaignErrorCount = 0;
-
         this.setCampaignForPlacement(placementId, campaign);
         this.handlePlacementState(placementId, PlacementState.READY);
     }
 
     private onNoFill(placementId: string) {
-        this._singleCampaignErrorCount = 0;
-
         this._nativeBridge.Sdk.logInfo('Unity Ads server returned no fill, no ads to show, for placement: ' + placementId);
         this.setCampaignForPlacement(placementId, undefined);
         this.handlePlacementState(placementId, PlacementState.NO_FILL);
     }
 
-    private onError(error: WebViewError | Error, placementIds: string[], rawAdPlan?: string) {
+    private onError(error: WebViewError | Error, placementIds: string[], session?: Session) {
         this.invalidateCampaigns(this._needsRefill, placementIds);
 
         if(error instanceof Error) {
             error = { 'message': error.message, 'name': error.name, 'stack': error.stack };
         }
 
-        Diagnostics.trigger('plc_request_failed', {
+        Diagnostics.trigger('auction_request_failed', {
             error: error,
-            rawAdResponse: rawAdPlan
-        });
+        }, session);
         this._nativeBridge.Sdk.logError(JSON.stringify(error));
 
         const minimumRefreshTimestamp = Date.now() + CampaignRefreshManager.ErrorRefillDelay * 1000;
@@ -176,24 +168,9 @@ export class CampaignRefreshManager {
         } else {
             this.setPlacementStates(PlacementState.NO_FILL, placementIds);
         }
-
-        if(this._singleCampaignMode) {
-            this._singleCampaignErrorCount++;
-
-            if(this._singleCampaignErrorCount === 1 && this._configuration.getAbGroup() === 5) {
-                const retryDelaySeconds: number = CampaignRefreshManager.QuickRefillTestDelay + Math.random() * CampaignRefreshManager.QuickRefillTestDelay;
-                this._nativeBridge.Sdk.logDebug('Unity Ads retrying failed campaign in ' + retryDelaySeconds + ' seconds');
-                this._refillTimestamp = Date.now() + CampaignRefreshManager.QuickRefillTestDelay * 1000;
-                setTimeout(() => {
-                    this._nativeBridge.Sdk.logDebug('Unity Ads retrying failed campaign now');
-                    this.refresh();
-                }, retryDelaySeconds * 1000);
-            }
-        }
     }
 
-    private onAdPlanReceived(refreshDelay: number, singleCampaignMode: boolean) {
-        this._singleCampaignMode = singleCampaignMode;
+    private onAdPlanReceived(refreshDelay: number) {
         if(refreshDelay > 0) {
             this._refillTimestamp = Date.now() + refreshDelay * 1000;
             this._nativeBridge.Sdk.logDebug('Unity Ads ad plan will expire in ' + refreshDelay + ' seconds');
@@ -214,11 +191,19 @@ export class CampaignRefreshManager {
                 this._nativeBridge.Sdk.logInfo('Unity Ads placement ' + placementId + ' status set to ' + PlacementState[placementState]);
                 this.setPlacementState(placementId, placementState);
                 this.sendPlacementStateChanges(placementId);
+                if(placementState === PlacementState.READY) {
+                    SdkStats.setReadyEventTimestamp(placementId);
+                    SdkStats.sendReadyEvent(placementId);
+                }
             });
         } else {
             this._nativeBridge.Sdk.logInfo('Unity Ads placement ' + placementId + ' status set to ' + PlacementState[placementState]);
             this.setPlacementState(placementId, placementState);
             this.sendPlacementStateChanges(placementId);
+            if(placementState === PlacementState.READY) {
+                SdkStats.setReadyEventTimestamp(placementId);
+                SdkStats.sendReadyEvent(placementId);
+            }
         }
     }
 }