+++ src/ts/Managers/OperativeEventManager.ts
@@ -65,7 +65,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.START)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.START]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve();
         }
 
@@ -91,7 +91,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.FIRST_QUARTILE)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.FIRST_QUARTILE]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
 
@@ -108,7 +108,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.MIDPOINT)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.MIDPOINT]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
 
@@ -125,7 +125,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.THIRD_QUARTILE)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.THIRD_QUARTILE]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
 
@@ -142,7 +142,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.SKIP)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.SKIP]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
         adUnit.getCampaign().getSession().setEventSent(EventType.SKIP);
@@ -179,7 +179,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.VIEW)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.VIEW]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
         adUnit.getCampaign().getSession().setEventSent(EventType.VIEW);
@@ -195,7 +195,7 @@ export class OperativeEventManager {
         if(adUnit.getCampaign().getSession().getEventSent(EventType.CLICK)) {
             Diagnostics.trigger('operative_event_not_sent', {
                 event: EventType[EventType.CLICK]
-            });
+            }, adUnit.getCampaign().getSession());
             return Promise.resolve(void(0));
         }
         adUnit.getCampaign().getSession().setEventSent(EventType.CLICK);
@@ -387,12 +387,17 @@ export class OperativeEventManager {
 
     private createClickEventUrl(adUnit: AbstractAdUnit<Campaign>): string {
         const campaign = adUnit.getCampaign();
-        let url: string;
+        let url: string | undefined;
         let parameters: any = {};
-        if((campaign instanceof PerformanceCampaign || campaign instanceof MRAIDCampaign) && campaign.getClickUrl()) {
-            url = campaign.getClickUrl();
-            parameters = { redirect: false };
-        } else {
+
+        if(campaign instanceof PerformanceCampaign || campaign instanceof MRAIDCampaign) {
+            const clickUrl = campaign.getClickUrl();
+            if(clickUrl) {
+                parameters = { redirect: false };
+                url = clickUrl;
+            }
+        }
+        if(!url) {
             url = [
                 OperativeEventManager.ClickEventBaseUrl,
                 campaign.getId(),