+++ src/ts/WebView.ts
@@ -38,6 +38,7 @@ import { StorageType } from 'Native/Api/Storage';
 import { FocusManager } from 'Managers/FocusManager';
 import { OperativeEventManager } from 'Managers/OperativeEventManager';
 import { Campaign } from 'Models/Campaign';
+import { SdkStats } from 'Utilities/SdkStats';
 
 import CreativeUrlConfiguration from 'json/CreativeUrlConfiguration.json';
 import CreativeUrlResponseAndroid from 'json/CreativeUrlResponseAndroid.json';
@@ -107,6 +108,7 @@ export class WebView {
 
             HttpKafka.setRequest(this._request);
             HttpKafka.setClientInfo(this._clientInfo);
+            SdkStats.setInitTimestamp();
 
             return this._deviceInfo.fetch();
         }).then(() => {
@@ -190,6 +192,9 @@ export class WebView {
             this._assetManager = new AssetManager(this._cache, this._configuration.getCacheMode(), this._deviceInfo);
             this._campaignManager = new CampaignManager(this._nativeBridge, this._configuration, this._assetManager, this._sessionManager, this._request, this._clientInfo, this._deviceInfo, this._metadataManager);
             this._campaignRefreshManager = new CampaignRefreshManager(this._nativeBridge, this._wakeUpManager, this._campaignManager, this._configuration);
+
+            SdkStats.initialize(this._nativeBridge, this._request, this._configuration, this._sessionManager, this._campaignManager, this._metadataManager);
+
             return this._campaignRefreshManager.refresh();
         }).then(() => {
             this._wakeUpManager.onNetworkConnected.subscribe(() => this.onNetworkConnected());
@@ -246,6 +251,8 @@ export class WebView {
             return;
         }
 
+        SdkStats.sendShowEvent(placementId);
+
         if(campaign.isExpired()) {
             this.showError(true, placementId, 'Campaign has expired');
             this._campaignRefreshManager.refresh();
@@ -254,7 +261,7 @@ export class WebView {
                 id: campaign.getId(),
                 willExpireAt: campaign.getWillExpireAt()
             });
-            Diagnostics.trigger('campaign_expired', error);
+            Diagnostics.trigger('campaign_expired', error, campaign.getSession());
             return;
         }
 
@@ -280,7 +287,7 @@ export class WebView {
                 const error = new DiagnosticError(new Error('No connection is available'), {
                     id: campaign.getId(),
                 });
-                Diagnostics.trigger('mraid_no_connection', error);
+                Diagnostics.trigger('mraid_no_connection', error, campaign.getSession());
                 return;
             }