+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# Thu Oct 19 20:45:59 UTC 2017
+
+* Trigger onStart when VPAIDAdUnit shows
+* Removes VPAIDAdUnit debug diagnostics
+
+# Wed Oct 18 21:19:54 UTC 2017
+
+* Add diagnostics to all VPAID third party events, success and failed
+* Stop quick retry test on parsing errors from group 5
+* Fix handling for non-integer screen width and height values
+
+# Mon Oct 16 20:19:29 UTC 2017
+
+* Adds ad response to all diagnostic messages
+
+# Thu Oct 12 20:04:06 UTC 2017
+
+* Fire VAST Impression on VPAID Ad Units
+* Fixes click URL construction for Operative Events
+
 # Wed Oct 11 01:48:02 UTC 2017
 
 * Fix typo in VPAID content type